Add files via upload
This is a creation of a Deep Learning Neural Network Model of 1-D Hyperelastic bar using the linear elastic method. It is a combination of Trapezoidal Integration and Deep Energy method to present a relationship between predicted and actual displacement of infinitely many points on the bar. The libraries used were NumPy, TensorFlow, matplotlib and time. This model basically uses Length of the Bar, Cross-sectional Area, Applied load, Young's Modulus and Number of Points on the bar to predict the displacemet. The overall results resemble the real life conditions near perfectly.